# coursetest.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/students": {
      "get": {
        "summary": "getAllStudents",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListStudent"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": "",
                      "studentId": "",
                      "name": "",
                      "major": "",
                      "grade": 0,
                      "email": "",
                      "createAt": ""
                    }
                  ],
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "createStudent",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": ""
              },
              "example": {
                "studentId": "45",
                "name": "姚静怡",
                "major": "exercitation est dolore cupidatat",
                "grade": 19,
                "email": "fvpts4_hkc23@yahoo.cn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudent"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": "",
                    "studentId": "",
                    "name": "",
                    "major": "",
                    "grade": 0,
                    "email": "",
                    "createAt": ""
                  },
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/students/{id}": {
      "get": {
        "summary": "getStudentById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "961ac0de-298b-4d60-96a8-dc2f0cf95610",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudent"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": "",
                    "studentId": "",
                    "name": "",
                    "major": "",
                    "grade": 0,
                    "email": "",
                    "createAt": ""
                  },
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateStudent",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "3b9dc8b8-a75a-4c3a-9a75-992724b2f979",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "student",
            "in": "query",
            "description": "",
            "required": false,
            "example": "{            \"id\": \"3b9dc8b8-a75a-4c3a-9a75-992724b2f979\",            \"studentId\": \"45\",            \"name\": \"姚静怡\",            \"major\": \"computer science\",            \"grade\": 19,            \"email\": \"aaa@bbb.com\",            \"createAt\": \"2025-10-26T20:45:48.5810944\"}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": ""
              },
              "example": {
                "id": "3b9dc8b8-a75a-4c3a-9a75-992724b2f979",
                "studentId": "45",
                "name": "姚静怡",
                "major": "computer science",
                "grade": 19,
                "email": "aaa@bbb.com",
                "createAt": "2025-10-26T20:45:48.5810944"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudent"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": "",
                    "studentId": "",
                    "name": "",
                    "major": "",
                    "grade": 0,
                    "email": "",
                    "createAt": ""
                  },
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteStudent",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "3b9dc8b8-a75a-4c3a-9a75-992724b2f979",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudent"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": "",
                    "studentId": "",
                    "name": "",
                    "major": "",
                    "grade": 0,
                    "email": "",
                    "createAt": ""
                  },
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/enrollments/student/{studentId}": {
      "get": {
        "summary": "getEnrollmentsByStudentId",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListEnrollment"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": "",
                      "studentId": "",
                      "courseId": "",
                      "createdAt": ""
                    }
                  ],
                  "timestamp": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "studentId": {
            "type": "string",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "major": {
            "type": "string",
            "description": ""
          },
          "grade": {
            "type": "integer",
            "description": ""
          },
          "email": {
            "type": "string",
            "description": ""
          },
          "createAt": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseListStudent": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Student",
              "description": "com.zjgsu.hx.schoolcoursesimple.model.Student"
            },
            "description": ""
          },
          "timestamp": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseStudent": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/Student",
            "description": ""
          },
          "timestamp": {
            "type": "string",
            "description": ""
          }
        }
      },
      "Enrollment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "studentId": {
            "type": "string",
            "description": ""
          },
          "courseId": {
            "type": "string",
            "description": ""
          },
          "createdAt": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseListEnrollment": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enrollment",
              "description": "com.zjgsu.hx.schoolcoursesimple.model.Enrollment"
            },
            "description": ""
          },
          "timestamp": {
            "type": "string",
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}